requirements.txt
pandas
PyYAML~=6.0.3
attrs==25.4.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.28.0
typing-extensions==4.15.0
